Move validDate and validTime to argumentUtil
