+++ app/util/time.js
@@ -1,7 +1,4 @@
 'use strict'
-const dateRegex = /(([0-2]?[0-9]|3[0-1])[-](0?[1-9]|1[0-2])[-][0-9]{4})/
-const timeRegex = /^(2[0-3]|[0-1]?[\d]):[0-5][\d]$/
-
 function diffDays (date1, date2) {
   const d1 = new Date(date1)
   const d2 = new Date(date2)
@@ -61,25 +58,6 @@ function isDst (date) {
   return Math.max(jan, jul) !== date.getTimezoneOffset()
 }
 
-function validDate (dateString) {
-  if (dateRegex.test(dateString)) {
-    const { day, month, year } = exports.getDateInfo(dateString)
-    const leapYear = year % 4 === 0
-    if (month === 0 || month === 2 || month === 4 || month === 6 || month === 7 || month === 9 || month === 11) {
-      return day <= 31
-    } else if (month === 3 || month === 5 || month === 8 || month === 10) {
-      return day <= 30
-    } else if (month === 1) {
-      return leapYear ? day <= 29 : day <= 28
-    }
-  }
-  return false
-}
-
-function validTime (timeString) {
-  return timeRegex.test(timeString)
-}
-
 module.exports = {
   diffDays,
   getDate,
@@ -87,7 +65,5 @@ module.exports = {
   getDurationString,
   getTime,
   getTimeInfo,
-  isDst,
-  validDate,
-  validTime
+  isDst
 }
